requirements.txt
# Core LLM & Vision
openai>=1.30.0

# Vector Database
weaviate-client>=4.5.0

# OCR
pytesseract>=0.3.10
Pillow>=10.3.0

# PDF processing
PyMuPDF>=1.24.0

# API framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Config & validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# RAGAS evaluation
ragas>=0.1.9
datasets>=2.19.0

# Utilities
python-dotenv>=1.0.1
